requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-auth==1.21.1
idna==2.10
kubernetes==11.0.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
urllib3==1.25.10
websocket-client==0.57.0
